feature: Changes the project to a Maven Plugin that executes Byte Buddy Maven Plugin on build time
This way, users don't need to directly use byte-buddy-maven-plugin and then
indicate an additional maven dependency that implements the actual byte code transformation.
Additionally, users are required to inform the fully qualified name of the class that implements a Byte Buddy Plugin
to perform the byte code transformation. It's too much setup.

Now, the user just needs to add a auto-class-accessors-maven-plugin, including an "apply" goal execution.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>